@@ -19,6 +19,10 @@ module BioportalWebUi
     config.i18n.available_locales = [:en, :fr, :it, :de]
     config.i18n.default_locale = :en
 
+    # Please, add to the `ignore` list any other `lib` subdirectories that do
+    # not contain `.rb` files, or that should not be reloaded or eager loaded.
+    # Common ones are `templates`, `generators`, or `middleware`, for example.
+    config.autoload_lib(ignore: %w[assets tasks])
 
     # Configuration for the application, engines, and railties goes here.
     #
@@ -29,8 +33,6 @@ module BioportalWebUi
     # config.eager_load_paths << Rails.root.join("extras")
 
     config.exceptions_app = self.routes
-
-
 
     config.settings = config_for :settings
     # Initialize configuration for KGCL change request functionality.
