@@ -1,8 +1,11 @@
 BioportalWebUi::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
+  # Include BioPortal-specific configuration options
+  require Rails.root.join('config', "bioportal_config_#{Rails.env}.rb")
+
   # Code is not reloaded between requests.
-  config.cache_classes = true
+  config.enable_reloading = false
 
   # Eager load code on boot. This eager loads most of Rails and
   # your application in memory, allowing both threaded web servers
@@ -74,12 +77,9 @@ BioportalWebUi::Application.configure do
   # Do not dump schema after migrations.
   config.active_record.dump_schema_after_migration = false
 
-  # Include the BioPortal-specific configuration options
-  require Rails.root.join('config', "bioportal_config_#{Rails.env}.rb")
-
   # Use a different cache store in the appliance.
-  config.cache_store = :mem_cache_store, ENV["MEMCACHE_SERVERS"] || "localhost:11211", { :namespace => 'bioportal_web_ui', :expires_in => 1.day }
-
+  config.cache_store = :mem_cache_store, ENV["MEMCACHE_SERVERS"] || "localhost:11211",
+                       { :namespace => 'bioportal_web_ui', :expires_in => 1.day }
 
   # Disable serving static files from the `/public` folder by default since
   # Apache or NGINX already handles this.
