@@ -3,31 +3,24 @@ require "active_support/core_ext/integer/time"
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
   config.assets.debug = true
+
   # In the development environment your application's code is reloaded any time
   # it changes. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
-  config.cache_classes = false
-  config.action_mailer.delivery_method = :letter_opener_web
+  config.enable_reloading = true
 
-  config.action_mailer.raise_delivery_errors = false
-
-  config.action_mailer.perform_deliveries = true
-  config.action_mailer.default_url_options = { host: '0.0.0:3000' } # Adjust the host/port as needed
-
-  config.action_mailer_letter_opener_location = Rails.root.join('tmp', 'my_mails')
-
-  
   # Do not eager load code on boot.
   config.eager_load = false
 
   # Show full error reports.
   config.consider_all_requests_local = true
 
-  # Enable server timing
+  # Enable server timing.
   config.server_timing = true
-  
-   # Allow all hosts in development
-   config.hosts = nil
+
+  # Allow all hosts in development.
+  config.hosts = nil
+
   # Enable/disable caching. By default caching is disabled.
   # Run rails dev:cache to toggle caching.
   if Rails.root.join("tmp/caching-dev.txt").exist?
@@ -35,9 +28,7 @@ Rails.application.configure do
     config.action_controller.enable_fragment_cache_logging = true
 
     config.cache_store = :memory_store
-    config.public_file_server.headers = {
-      "Cache-Control" => "public, max-age=#{2.days.to_i}"
-    }
+    config.public_file_server.headers = { "Cache-Control" => "public, max-age=#{2.days.to_i}" }
   else
     config.action_controller.perform_caching = false
 
@@ -50,7 +41,19 @@ Rails.application.configure do
   # Don't care if the mailer can't send.
   config.action_mailer.raise_delivery_errors = false
 
+  # Disable caching for Action Mailer templates even if Action Controller
+  # caching is enabled.
   config.action_mailer.perform_caching = false
+
+  config.action_mailer.default_url_options = { host: "localhost", port: 3000 }
+
+  # Tells Rails to use the letter_opener_web gem to “deliver” emails in development by opening them in the browser,
+  # rather than actually sending them.
+  config.action_mailer.delivery_method = :letter_opener_web
+
+  # Tells the letter_opener gem to save the generated email HTML files to a specific directory — in this case,
+  # tmp/my_mails — rather than the default tmp/letter_opener.
+  config.action_mailer_letter_opener_location = Rails.root.join('tmp', 'my_mails')
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
@@ -66,6 +69,9 @@ Rails.application.configure do
 
   # Highlight code that triggered database queries in logs.
   config.active_record.verbose_query_logs = true
+
+  # Highlight code that enqueued background job in logs.
+  config.active_job.verbose_enqueue_logs = true
 
   # Suppress logger output for asset requests.
   config.assets.quiet = true
@@ -89,8 +95,14 @@ Rails.application.configure do
   # config.i18n.raise_on_missing_translations = true
 
   # Annotate rendered view with file names.
-  # config.action_view.annotate_rendered_view_with_filenames = true
+  config.action_view.annotate_rendered_view_with_filenames = false
 
   # Uncomment if you wish to allow Action Cable access from any origin.
   # config.action_cable.disable_request_forgery_protection = true
+
+  # Raise error when a before_action's only/except options reference missing actions.
+  config.action_controller.raise_on_missing_callback_actions = true
+
+  # Apply autocorrection by RuboCop to files generated by `bin/rails generate`.
+  # config.generators.apply_rubocop_autocorrect_after_generate!
 end
