@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_11_30_210229) do
+ActiveRecord::Schema[7.2].define(version: 2023_11_30_210229) do
   create_table "analytics", id: :integer, charset: "utf8mb4", force: :cascade do |t|
     t.string "segment"
     t.string "action"
@@ -50,5 +50,4 @@ ActiveRecord::Schema[7.0].define(version: 2023_11_30_210229) do
     t.datetime "created_at", precision: nil
     t.datetime "updated_at", precision: nil
   end
-
 end
